requirements.txt
Flask==3.1.0
python-decouple==3.8
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
selenium==4.27.1
webdriver_manager==4.0.2
